test: Phase 0 of v2 testing suite overhaul
- add docs/testing-plan.md: categories, phases, and conventions for
  the v2 testing suite (see PR/plan section 5 for the build order)
- resolve the stale subscribe test per the ratified body-less
  subscribe semantics (plan section 7): replace the unreachable 403
  expectation with test_subscribe_ignores_supplied_member_address
- dedupe the deployment_dir/backend fixtures and JWT env preamble
  (previously copy-pasted across four files) into tests/conftest.py
- register unit/integration/contract/e2e markers, auto-applied from
  the tests/ subdirectory; exclude e2e from default runs
- scope pytest-cov to the four v2 packages (baseline: 41% line)
- add respx dev dependency for upcoming httpx-mocking tests
