+++ pyproject.toml
@@ -84,6 +84,19 @@ exclude = [
 [tool.mypy]
 ignore_missing_imports=true
 
+[tool.coverage.run]
+# Scope coverage to the v2 packages; run with `uv run pytest --cov`.
+source_pkgs = [
+    "mail_protocol",
+    "mail_server",
+    "mail_client",
+    "mail_daemon",
+]
+
+[tool.coverage.report]
+show_missing = true
+skip_covered = true
+
 [dependency-groups]
 dev = [
     "mypy>=1.17.1",
@@ -91,6 +104,7 @@ dev = [
     "pytest-asyncio>=1.1.0",
     "pytest-cov>=7.0.0",
     "referencing>=0.36.2",
+    "respx>=0.23.1",
     "ruff>=0.12.12",
     "tqdm>=4.67.1",
 ]
