+++ pytest.ini
@@ -1,6 +1,7 @@
 [pytest]
-addopts = -ra -p no:cacheprovider
-# Default collection is the active MAIL v2 suite.
+addopts = -ra -p no:cacheprovider -m "not e2e"
+# Default collection is the active MAIL v2 suite (e2e excluded; run
+# `uv run pytest -m e2e` for the full-system subprocess tests).
 # Run `uv run pytest src/mail/legacy/tests` for archived MAIL v1 tests.
 testpaths = tests
 python_files = test_*.py
@@ -9,3 +10,8 @@ python_functions = test_*
 filterwarnings =
     ignore::DeprecationWarning
 asyncio_mode = auto
+markers =
+    unit: pure-logic tests (tests/unit); no network, no app object
+    integration: in-process FastAPI app tests with real auth (tests/integration)
+    contract: spec/OpenAPI conformance tests (tests/contract)
+    e2e: full-system subprocess tests (tests/e2e); excluded from the default run
